+++ src/BreakpointExtension.php
@@ -20,25 +20,22 @@ use Twig\TwigFunction;
  */
 final class BreakpointExtension extends AbstractExtension
 {
-    public function getName()
+    public function getName(): string
     {
         return 'breakpoint';
     }
 
-    public function getFunctions()
+    /**
+     * @return TwigFunction[]
+     */
+    public function getFunctions(): array
     {
         return [
             new TwigFunction('breakpoint', [$this, 'setBreakpoint'], ['needs_environment' => true, 'needs_context' => true]),
         ];
     }
 
-    /**
-     * If Xdebug is detected, makes the debugger break.
-     *
-     * @param Environment $environment the environment instance
-     * @param mixed       $context     variables from the Twig template
-     */
-    public function setBreakpoint(Environment $environment, $context)
+    public function setBreakpoint(Environment $environment, $context): void
     {
         if (function_exists('xdebug_break')) {
             $arguments = func_get_args();